Add files via upload
The main objective of this project is to ease the timetable generation process in institutions. To generate time table automatically for individual classes and staffs. To view timetable in easier manner. To increase the efficiency of the work Software Requirements • The software used to build this project are : • PYTHON • HTML • CSS • JAVASCRIPT • The project can run on : • Windows • Linux 64-bit; Linux 32-bit • macOS